feat: set-bacenta-status
